Split utils to align with Unreal; Removed unused imports; Added missing docstrings; Removed camel case function and var names
