+++ resource/plugins/python/loader/importers/nuke_sequence_loader_importer.py
@@ -9,7 +9,7 @@ import ftrack_api
 import nuke
 
 from ftrack_connect_pipeline_nuke import plugin
-from ftrack_connect_pipeline_nuke.utils import custom_commands as nuke_utils
+from ftrack_connect_pipeline_nuke import utils as nuke_utils
 from ftrack_connect_pipeline_nuke.constants import asset as asset_const
