+++ resource/plugins/python/opener/finalizers/nuke_default_opener_finalizer.py.orig
@@ -4,7 +4,7 @@
 import ftrack_api
 
 from ftrack_connect_pipeline_nuke import plugin
-from ftrack_connect_pipeline_nuke.utils import custom_commands as nuke_utils
+from ftrack_connect_pipeline_nuke import utils as nuke_utils
 
 
 class NukeDefaultOpenerFinalizerPlugin(plugin.NukeOpenerFinalizerPlugin):
