+++ resource/plugins/python/publisher/exporters/nuke_camera_publisher_exporter.py
@@ -7,7 +7,7 @@ import clique
 import tempfile
 
 from ftrack_connect_pipeline_nuke import plugin
-from ftrack_connect_pipeline_nuke.utils import custom_commands as nuke_utils
+from ftrack_connect_pipeline_nuke import utils as nuke_utils
 
 import nuke
