+++ resource/plugins/python/publisher/exporters/nuke_reviewable_publisher_exporter.py.orig
@@ -9,7 +9,7 @@ import ftrack_api
 import nuke
 
 from ftrack_connect_pipeline_nuke import plugin
-from ftrack_connect_pipeline_nuke.utils import custom_commands as nuke_utils
+from ftrack_connect_pipeline_nuke import utils as nuke_utils
 
 
 class NukeReviewablePublisherExporterPlugin(
@@ -26,7 +26,7 @@ class NukeReviewablePublisherExporterPlugin(
         node_name = collected_objects[0]
         input_node = nuke.toNode(node_name)
         selected_nodes = nuke.selectedNodes()
-        nuke_utils.cleanSelection()
+        nuke_utils.clean_selection()
 
         try:
             mode = (options.get('mode') or 'render').lower()
@@ -157,7 +157,7 @@ class NukeReviewablePublisherExporterPlugin(
 
         finally:
             # restore selection
-            nuke_utils.cleanSelection()
+            nuke_utils.clean_selection()
             for node in selected_nodes:
                 node['selected'].setValue(True)
