+++ resource/plugins/python/publisher/exporters/nuke_sequence_publisher_exporter.py
@@ -7,7 +7,7 @@ import clique
 import tempfile
 
 from ftrack_connect_pipeline_nuke import plugin
-from ftrack_connect_pipeline_nuke.utils import custom_commands as nuke_utils
+from ftrack_connect_pipeline_nuke import utils as nuke_utils
 
 import nuke
 import shutil
@@ -28,7 +28,7 @@ class NukeSequencePublisherExporterPlugin(plugin.NukePublisherExporterPlugin):
         node_name = collected_objects[0]
         input_node = nuke.toNode(node_name)
         selected_nodes = nuke.selectedNodes()
-        nuke_utils.cleanSelection()
+        nuke_utils.clean_selection()
         try:
             mode = (options.get('mode') or 'render').lower()
             if mode == 'render':
@@ -197,7 +197,7 @@ class NukeSequencePublisherExporterPlugin(plugin.NukePublisherExporterPlugin):
                 )
         finally:
             # restore selection
-            nuke_utils.cleanSelection()
+            nuke_utils.clean_selection()
             for node in selected_nodes:
                 node['selected'].setValue(True)
