+++ resource/plugins/python/publisher/validators/nuke_file_saved_publisher_validator.py.orig
@@ -6,7 +6,7 @@ import ftrack_api
 import nuke
 
 from ftrack_connect_pipeline_nuke import plugin
-from ftrack_connect_pipeline_nuke.utils import custom_commands as nuke_utils
+from ftrack_connect_pipeline_nuke import utils as nuke_utils
 
 
 class NukeFileSavedPublisherValidatorPlugin(
