+++ source/ftrack_connect_pipeline_nuke/client/load.py
@@ -9,7 +9,6 @@ from ftrack_connect_pipeline_qt.client import load
 import ftrack_connect_pipeline.constants as constants
 import ftrack_connect_pipeline_qt.constants as qt_constants
 import ftrack_connect_pipeline_nuke.constants as nuke_constants
-from ftrack_connect_pipeline_nuke.utils.custom_commands import get_main_window
 
 
 class NukeQtAssemblerClientWidget(load.QtAssemblerClientWidget):
