+++ source/ftrack_connect_pipeline_nuke/constants/asset/modes.py
@@ -1,7 +1,7 @@
 # :coding: utf-8
 # :copyright: Copyright (c) 2014-2022 ftrack
 
-from ftrack_connect_pipeline_nuke.utils import custom_commands as nuke_utils
+from ftrack_connect_pipeline_nuke import utils as nuke_utils
 
 # Load Modes
 IMPORT_MODE = 'import'
