+++ source/ftrack_connect_pipeline_nuke/plugin/__init__.py
@@ -5,7 +5,7 @@ import nuke
 from ftrack_connect_pipeline import plugin
 from ftrack_connect_pipeline_qt import plugin as pluginWidget
 from ftrack_connect_pipeline_nuke import constants as nuke_constants
-from ftrack_connect_pipeline_nuke.utils import custom_commands as nuke_utils
+from ftrack_connect_pipeline_nuke import utils as nuke_utils
 from ftrack_connect_pipeline_nuke.asset import NukeFtrackObjectManager
 from ftrack_connect_pipeline_nuke.asset import NukeDccObject
