+++ source/ftrack_connect_pipeline_nuke/plugin/open/importer.py
@@ -14,7 +14,7 @@ from ftrack_connect_pipeline_nuke.plugin import (
 
 from ftrack_connect_pipeline_nuke.constants import asset as asset_const
 from ftrack_connect_pipeline_nuke.constants.asset import modes as load_const
-from ftrack_connect_pipeline_nuke.utils import custom_commands as nuke_utils
+from ftrack_connect_pipeline_nuke import utils as nuke_utils
 
 
 class NukeOpenerImporterPlugin(plugin.OpenerImporterPlugin, NukeBasePlugin):
