+++ source/ftrack_connect_pipeline_nuke/plugin/publish/finalizer.py
@@ -13,7 +13,7 @@ from ftrack_connect_pipeline_nuke.plugin import (
     NukeBasePluginWidget,
 )
 from ftrack_connect_pipeline_nuke.constants import asset as asset_const
-from ftrack_connect_pipeline_nuke.utils import custom_commands as nuke_utils
+from ftrack_connect_pipeline_nuke import utils as nuke_utils
 
 
 class NukePublisherFinalizerPlugin(
